+++ lib/api.dart
@@ -10,6 +10,8 @@ export 'src/api/api_endpoint.dart';
 export 'src/api/api_request_method.dart';
 export 'src/api/route.dart';
 
+export 'src/api/authentication/jwt.dart';
+
 export 'src/api/hub/api_resource.dart';
 export 'src/api/hub/static_list_api_resource.dart';